some exports cleanup
still a WIP to see if I can continue to attack circular architecture
