+++ src/index.ts
@@ -1,10 +1,8 @@
 export * from './api';
-export { default as identityProvider } from './entity-idp';
-export { default as serviceProvider } from './entity-sp';
-export * as extractor from './extractor';
-export { default as libsaml } from './libsaml';
-export { default as metadataIdp } from './metadata-idp';
-export { default as metadataSp } from './metadata-sp';
-export * from './types';
+export * from './entity-idp';
+export * from './entity-sp';
+export * from './error';
+export * from './libsaml';
+export * from './metadata-idp';
+export * from './metadata-sp';
 export * from './urn';
-export * as utility from './utility';
